requirements.txt
flask
gradio
tensorflow
numpy
pillow
